Refactored Again
Whoa boy what a train ride this has been (literally). I refactored everything for general domains where we can define the appropriate (discretized) functions.

It still fully works on the grid, but there are some things to double check. (ie, delusional math mistakes on the traain)

I defined an abstract class called DiffeoOTDomain that lays out the functions we must define on a space. We have already done this for S^1, so that should be easy enough to port over. Have fun @Mao. Also, most of the work has been done on meshes, I just need to get a HodgeStar implemented and then we can do OT on the sphere (or any mesh for that matter)